selesai modul php mvc

+++ app/views/mahasiswa/index.php
@@ -1,47 +1,61 @@
 <div class = "container mt-5">
 
-<div class="row">
-<div class="col-lg-6">
-  <?php Flasher::flash();?></div>
-</div>
+  <div class="row">
+  <div class="col-lg-6">
+    <?php Flasher::flash();?></div>
+  </div>
 
 
+  <!-- Pencarian -->
+  <div class="row mb-3">
+    <div class="col-lg-6">
+
+      <form action = "<?=BASEURL; ?>/mahasiswa/cari" method = "post">
+        <div class="input-group">
+          <input type="text" class="form-control" placeholder="cari mahasiswa" name="keyword" id="keyword" autocomplete = "off" >
+          <div class="input-group-append">
+            <button class="btn btn-primary" type="submit" id="button-cari">Cari</button>
+          </div>
+        </div>
+      </form>
+
+    </div>
+  </div>
+
+
+  <!-- tombol tambah -->
+  <div class="row mb-4">
+    <div class="col-lg-6">
+      <!-- Button trigger modal -->
+      <button type="button" class="btn btn-primary tombolTambahData" data-toggle="modal" data-target="#exampleModalLong">
+        Tambah Data
+      </button>
+    </div>
+  </div>
   
-        <div class="col-lg-6">
-            <h3>Daftar Mahasiswa</h3>
-            <ul class="list-grup mt-5">
-            
-            <?php foreach( $data['mhs'] as $mhs ) : ?>
+  <div class="row">
+    <div class="col-lg-6">
 
-            <li class="list-group-item ">
-
+      <h3>Daftar Mahasiswa</h3>
+      <ul class="list-group">
+          <?php foreach( $data['mhs'] as $mhs ) : ?>
+          <li class="list-group-item ">
             <?= $mhs['nama'];?>
-            <a href="<?=BASEURL;?>/mahasiswa/hapus/<?=$mhs['id'];?>" class="badge badge-danger float-right ml-2"
-            onclick="return confirm('yakin');">hapus</a>
-            
+            <a href="<?=BASEURL;?>/mahasiswa/hapus/<?=$mhs['id'];?>" class="badge badge-danger float-right ml-2"onclick="return confirm('yakin');">hapus</a>
             <a href="<?=BASEURL;?>/mahasiswa/ubah/<?=$mhs['id'];?>" class="badge badge-success float-right  ml-2 tampilModalUbah" data-toggle="modal"  data-target="#exampleModalLong" data-id="<?=$mhs['id'];?>">ubah</a>
             <a href="<?=BASEURL;?>/mahasiswa/detail/<?=$mhs['id'];?>" class="badge badge-primary float-right">detail</a>
-            
-            </li>
-
-            <?php endforeach; ?>
-            </ul>
-            
-        </div>
-    </>
+          </li>
+        <?php endforeach; ?>
+      </ul> 
+    </div>
+  </div>
+    
 
 
 
 
 </div>
 
-
-<div class="container">
-<!-- Button trigger modal -->
-<button type="button" class="btn btn-primary tombolTambahData" data-toggle="modal" data-target="#exampleModalLong">
-  Tambah Data
-</button>
-</div>
 
 <!-- Modal -->
 <div class="modal fade" id="exampleModalLong" tabindex="-1" role="dialog" aria-labelledby="exampleModalLongTitle" aria-hidden="true">
